Create a central "content-container" tailwind utility class in scaffold/tailwind.config.js
