+++ packages/ui/tailwind.config.js
@@ -1,7 +1,1 @@
-const theme = require("../scaffold/tailwind.theme");
-
-module.exports = {
-  content: ["./src/**/*.{ts,tsx,html}", "./index.html"],
-  theme,
-  plugins: [],
-};
+module.exports = require("../scaffold/tailwind.config.js");
